Allow all users to see the grails-errorhandler
This fixes the exception thrown when logging in. The reason for this
exception was that with an ActiveDirectory there is an expected exception
which is ignored by the framework.

Nevertheless the exception triggers the error handling which is on
/grails-errorhandler/ but at that point the user is not yet logged in, so
she gets redirected to the login page again. The original rederict URL
gets lost by that and once the user is properly logged in she gets
redirected to the redirect URL, which is now grails-errorhandler and
shows ther error500 page. In this case the exception attribute of the
request is not set, which has been the major indication that there is not
a real error at all.

By allowing all users to access grails-errorhandler the user does not
get redirected to the login page again and by that gets properly
redirected to the requested page.