Add links to fastqc files in Sequence list
To be able to add the links towards Fastqc files from the Sequence
datatable a new service method is added which returns all DataFiles for a
given list of Sequences.

Inside the controller this retrieved list is merged with the list of all
Sequences and the DataFiles for one Sequence are added as an additional
property to the JSON return string.

Inside the JavaScript part links are generated for the DataFiles if
present and added to the FastQC info field.