Introduce helper methods to createLinks and link markup
The idea behind this change is to be more DRY in the JavaScript part.
Lots of code created again and again links to controller actions.

This is now replaced by a method following the same approach as the
groovy tag to create a link:
$.otp.createLink({
    controller: 'foo',
    action: 'bar',
    id: 12345,
    parameters: {
        foobar: 'foobarbaz'
    }
})

Another advantage is that this makes the code much more readable
especially when creating the HTML markup:
$.otp.createLinkMarkup({
    controller: 'foo',
    action: 'bar',
    id: 12345,
    parameters: {
        foobar: 'foobarbaz'
    },
    text: "Awesome link text"
})

REVIEW: 184