Fix bug in /individual/insert saving at wrong button
As no more existing function was called when user clicked
on plus button to get a new input field for SampleIdentifier
on at least second Sample it was tried to save the page.
This led to an error if not all required fields were
filled and otherwise to unexpected behaviour.