Move code to modify coverage file to separate jar
Currently, the code to
- map chromosomes names
- filter out chromosomes
- sort chromosomes
of the coverage file are placed directly in OTP.

But OTP is not allowed to write to the lsdf:
- first, the lsdf is mounted readonly
- second, the directory structure is owned by [redacted], who is the only
user allowed to write data in the directories.

Also for the sorting it is necessary to load the complete file into the
memory to handle it. The coverage file is around 50MB big,
so it is no good idea to hold them in the server memory, specially,
if multiple files are handled parallel.

For this reasons we decide to do the work in a separate jar file,
and provide the information about mapping, filtering and sorting in a
file.

Therefore the following things are done:
- create new project coveragePlotMFS for do the mapping,
  filtering and sorting
- move functionality from OTP to new project
- adapt functionality, if needed
- add needed service method
- delete not longer needed service method