OTP-585: Bugfix for transfer workflow
The transfer workflow failed for projects on the other LSDF due to the
setup of the infrastructure. We now force processing on the DKFZ where this
is necessary. In the copy step, we copy data from one LSDF to the other LSDF and move
the files to the right place from within the other cluster.

This commit also includes a bug fix for the run IDs in the statistics file
(OTP-592).