OTP-684: Hide mapping for MMML for normal users
In MMML we have a mapping from intern to extern ids. 
The extern ids are created by us, so we get the data with the intern
ids. This mapping is visible in OTP, which was okay till yet. But now
this is not longer okay. Therefore this need to be changed.