OTP-754: abstract class hierarchy problem: solution 2
For unknown reason grails/groovy has problems with abstract class hierarchy containing final methods.

After a grails clean, the first grails run-app works fine, but the second grails run-app produce the following error:
Error Fatal error during compilation org.apache.tools.ant.BuildException: java.lang.VerifyError:
class de.dkfz.tbi.otp.job.processing.AbstractEndStateAwareJobImpl overrides final method start.()V (Use --stacktrace to see the full trace)

The method name can be differ, but the result is the same. But the method is only defined in AbstractJobImpl,
not in AbstractEndStateAwareJobImpl, so it shouldn't happen.

Looking in the history, there a similar problem is discribed by Martin in 51dc3db.

To solve the problem, the final methods in AbstractJobImpl are changed to not final, so the can be overloaded.
The first try to make AbstractEndStateAwareJobImpl not abstract do not work, because it crashes grails test-app all the time.