BUGFIX: The method inferKitInformationForOldLaneFromNewLane fails for runsegments containing not fastq files.
The current implementation assumes, that a run segment contains only
fastq files. But that is not always true, there are projects we getting
usually also bam files. And bam files are not part of a seqTrack.
Therefore, the null pointer introduce by such a data file needs to be
filtered out, so that the next check do not fail.