OTP-759: add database indexes to dom classes to speed up workflow overview
The following indexes speed up the queries used for the page "processes/list"
Before it takes around 15 secs to get the information, after it needs only around 2 secs.
The decision to this indexes are come from the queries used for getting the data for the page

The indexes are defined in the DOM via gorm mapping. Also a SQL script to create the indexes are provided.