OTP-816: ProcessingSteps are not created transactionally
For some reason the transaction was disabled for the SchedulerService.
The history (6c7f3154bc5fcc52323db97d2b1475fceb66d1e3) says, it was
because of lazy initializing exception.

But disabling of transaction makes problems, because then only some
objects are created and not all in case of failure, which result in an
inkonsistent database. Also in the other methods no transaction will
make problems.

Therefore I activate the transaction again.

One cylclic dependency which was introduced between I have also resolved.