Revert "OTP-816: ProcessingSteps are not created transactionally"
This reverts the commit commit 199a5f6179f026c3e6721039309acb6c8768cbfa.

The revert is needed, because with transaction in the doEndCheck of the SchedulerService
the creation of the next processing step always produces the following error:
Caused by: org.hibernate.StaleObjectStateException: Row was updated or deleted by another transaction (or unsaved-value mapping was incorrect): [de.dkfz.tbi.otp.job.processing.RestartedProcessingStep#10195597]
	at de.dkfz.tbi.otp.job.scheduler.SchedulerService.createProcessingStep(SchedulerService.groovy:823)
	at de.dkfz.tbi.otp.job.scheduler.SchedulerService.createNextProcessingStep(SchedulerService.groovy:221)
	at de.dkfz.tbi.otp.job.scheduler.SchedulerService.doEndCheck(SchedulerService.groovy:546)
	at de.dkfz.tbi.otp.job.scheduler.Scheduler.doEndCheck(Scheduler.groovy:164)
	at de.dkfz.tbi.otp.job.scheduler.SchedulerService$_schedule_closure6.doCall(SchedulerService.groovy:308)
	at grails.plugin.executor.PersistenceContextRunnableWrapper$_run_closure1.doCall(PersistenceContextRunnableWrapper.groovy:34)
	at grails.plugin.executor.PersistenceContextWrapper.wrap(PersistenceContextWrapper.groovy:35)

As fast fix the problematic commit is reverted.
But no transaction are also problematic, so another solution needs to be found.