BUGFIX: handle null values on run page and trim headers of metadata files
If a metadata file has no SAMPLE_ID column, the run pages of the corresponding run
throws a nullpointer exception in the GSP code. Detecting the place is difficult,
since the stack trace shows the line number of the generated java code. To avoid this
in future, I change the code call to be null safe.

Also it is confusing, that we trim meta data values, but not the header values
of a metadatafile. This is changed also, so both are corresponding.