OTP-1089: allow reference genome per sample type
Currently we can define Reference Genome only per Project and SeqType.
But in some Projects we need for some SampleTypes (for example for
xenograft) another Reference Genome then for the others SampleTypes.

Therefore we need the possibility to define a ReferenceGenome depending
on Project, SeqType and SampleType. But because that are very much
combination and in many Projects we do not need SampleType depending
configuration we decided to use a hierarchy definition.

Therefore it should be possible to define Reference Genome per:
- Project and SeqType
- Project, SeqType and SampleType

Since we know, that Xenografts SampleTypes have always a specific
Reference Genome we decide further to have a mechanisms to say that
some SampleTypes have always a specific reference genome.

- adapt domains
  - add SampleType to ReferenceGenomeSeqType
  - add enum to SampleType to distinct the cases
  - adapt SeqTrack.getConfiguredReferenceGenome to handle new functionality
- adapt services
  - remove referenceGenomeService.referenceGenome and adapt all occurrences
  - adapt OverallQualityAssessmentMergedService to use the used reference genome
- update wiki script "Store reference genome - project"
- adapt sample swap script: because the check of not changed reference
genome during movement between projects can not done any-more on project
seqtype level the code is changed to do always a realignment. On the way
some code duplication and inconsistencies are cleaned up.