OTP-1422: RoddyBamFile
* updated AbstractBamFile with new abstract methods
* introduced AbstractMergedBamFile
** numberOfMergedLanes is not nullable
* make ProcessedMergedBamFile extend AbstractMergedBamFile
* introduce RoddyBamFile
* refactored some old tests
* moved bamFile methods into AbstractBamFile

Initially we tried to introduce contraint:
numberOfMergedLanes == containedSeqTracks.size()
on the level of AbstractMergedBamFile. Because it was too much
work to fix all the old tests, we decided to have such contraint
on the level of RoddyBamFile. The already improved tests are kept.

ProcessedMergedBamFileService.updateNumberOfMergedLanes has been
deleted, because the numberOfMergedLanes is set at the time when
the processedMergeBamFile is created. Corresponding code has been
clean up.