OTP-1528: Provide SampleType in Roddy-PanCan call
the panCan-Wf needs to know the sample type of the bam file.
in roddy there is hard coded list with all the different sample types.
In case a new sample types comes up it has to be inserted in the list.
To prevent the case that jobs fail because the sample type is not known we decided to provide the sample type in the roddy call itself