OTP-1451 prework: add property statSizeFileName to ReferenceGenomeProjectSeqType
Roddy need not only the directory to the stat file, but also the file
name itself. Since a reference genome can have multiple such files, the
file name need to be stored in OTP. Different Project can use different
files, therefore the file name can not saved in the reference genome
table. Therefore it was decided to store it in ReferenceGenomeSeqtype,
so for every project and seq type an own name is stored. The MergingWorkpackage
store the value used for the proccessing based on this MergingWorkpackage.