OTP-1451 prework: add correction of permissions
Since there are problems with the permission set through roddy, OTP need to adapt permission.

For unknown reason, the configured file permissions are only used for roddy own files,
but not for the result files, which have always 600.

Additional the permissions of directoresy are set to 660. This permissions are required
before we move file to the final directory, but after moving the moved directories should
have the permissions 750.

Therefore a script is created, which correct the permissions of directories
depending before and after moving of the files. This script is called
before and after moving to have always the correct permissions.

Since it needs to be executed remote through the [redacted] user,
a wrapper script is also created.

The permissions are set as following:
- directories inside tmp roddy directory: 2770
- directories outside tmp roddy directory: 2750
- files except bam files: 640
- bam files: 644

Additional the group of the tmp roddy directory is set to localGroup, so that only our group has
access to this directory and not normal users.