OTP-1660: Fix problem to update SnvConfig to new Version
A bug in the constraint requere, that ConfigPerProject always refers to
a not deprecated ExternalScript, also if ConfigPerProject is obsolate.
That cause that it is not possible to adapt the script version.

For that reason the constraint is adapted, that obsolate
ConfigPerProject can also reference to an depricated snv version.