Bugfix: Move first 'Correct file permission' call to ExecutePanCanJob.validate
In Production the file check canRead return false for roddy files with permission 600, but in our test system true.
We don't know, why the bavior is different, it was not possible to reproduce it local with workflow test.

Nevertheless the bahavior in productio is the correct. Therefore we decide to correct the file permission before we
check files to be readable