Use File.exist() instead of waitUntilExists for deleteResultFileIfExists
Change check for already existing result files to the method File.exists() to save much time.
Because normally there is no file and then it waits the full waiting time of currently a minute
till it decide that the file does not exist. In case there is a file, it was created before the job
was restarted, and in that case the file will be there without any waiting.

Therefore the check was changed.