Fix BedUtils creation
The Tool BedUtils were configured to include all depending classes in the created jar file.
As long as the version of the dependency are the same as in the project usses BedUtils it was no Problem.
But now a newer version of groovy and spring is used in OTP, so there are conflicting files.

Therefore the build script is updated to not include the dependency anymore.