Fix command to delete [redacted] directory content
The command ExecuteRoddyCommandService.deleteContentOfOtpRoddyDirectory should delete the
content of directories owned by [redacted].

The old script asumes that it is called for a directory own by [redacted], which is not always true.
In that case the deletion failed and the method throw an exception.

The adaption search for non empty directoriries owned by [redacted] and call rm for the
content of these directories.