Fix delay problem in PbsMonitor
Currently the pbs monitor fetch first all cluster ids from pbs and
create then a copy of the cluster ids to check. But if in the period
between fetching cluster jobs from pbs and creating the copy of the
list of known cluster jobs new cluster job are added, they are not
in the pbs output and therefore OTP think they are finished.

To solve the Problem the copy is created before the pbs server is
asked, so that no new cluster job can be added for this check round.