Bugfix: Remove uneccesary synchronised in FastqcJob which cause deadlocks
The synchronised block cause a deadlock together with the database, if:
- A thread locks a database object before it reach the synchronized block.
- Another thread comes in the synchronized block and wants to modify an object where the first thread has lock.

This cause an deadlock, since the first thread wait that the second thread lives the synchronized block,
and the second thread wait that the first thread release the database lock.

Remove the synchronisation in OTP moves the synchronisation problem complete in the database,
so postgres can handle that problem.

See also OTP-1629, which is caused by the same problem, but on another place.

When this synchronisation was introduced, OTP saves the complete fastq file in the database. If multiple files
was uploaded parallel, it slow down OTP dramatically.
But since OTP saves now only some values, the synchronization is not needed anymore.