Adapt SampleIdentifierService to create SampleTypes with expected specificReferenceGenome value
Currently the autoimport fail, if a new sample type is created and the sample should be aligned, because
the value specificReferenceGenome is not set for alignment. Since the value depends currently only on xenograft,
it is included in the create process, so the auto import do not fail if ne sample types are created.