Bugfix: fix problem in RNA paired alignment workflow test
In the workflow test always a segmention fault occurs for featureCount of the rna alignment workflow.
Since the Problem occur only for tests, it seems to have something to do with the test data.
After some work of Navveed we found out, that reducing the thread count to one for this command resolve the Problem.

Therefore Naveed create a new plugin version which the possibility to set the value via cvalue.

Since the Proble occur only for the test data, we set it only there to one.