Improvement: Add check to the copy fastqc command
Currently the copy and md5sum script is executed, but the return value is not checked.
Problems during copy are only found, if the expected file does not exist. But if the md5sum
is not correct, it is not found at all.

This change the code to check the output of the script and activate to fail for the first error.