Imrovement: Add fastqc blacklist for query running workflows in ProcessingProgressMonitor
Currently there are many fastqc workflows are failed because of data problems and it seems that they are not solved.
Also there are Problems of the index files, which is not supported by OTP.

All this errors are added to an blacklist and will now ignored to detect not finished fastqc workflows.

For other ways of select seqTracks the blacklist is not used.