Migrate Roddy configs to the new plugin version
Sinse the plugin version is part of the xml (file name and identifier in file),
it is also necessary to update the config files.

Therefore the script to migrate binaries to module was adapted to also update the plugin version.
Also the script was extended to handle also all non pancan configs.

Additional, a bug was fixed in the migration of project files.

The templates are also adapt to use always double quotes instead of single quotes.