OTP-2801: Retry execution a second time, if a 'channel is not opened' error occurred
Sometime the command execution fail with the exception 'channel is not opened'. That indicates,
that the session was closed and needs to be recreated. Therefore it is necessary to wait some time and
reexecute the command. The code change automized these steps.

In case it failes again, it seems to be a problem which are not solved by waiting some time,
therefore it should not try again.