OTP-2723: stop doing fastQC for withdrawn SeqTracks.
By checking if any of their datafiles is withdrawn.
This SQL implementation matches the (current) implementation of
SeqTrack.isWithdrawn()

We also stop using deprecated SeqTrackService.alignableSeqtypes, and
use replacement method SeqType.getAllAlignableSeqTypes() and/or
getAllAnalysableSeqTypes instead.
This changes the meaning, though not always the behaviour, of OTP,
and thus required test adaptations. Unfortunately, this was an old-style,
deeply tangled 6-in-1 "unit"-test... We adapted/rewrote all tests for
getSeqTrackReadyForFastqcProcessing to Spock.