new: OTP-2978: Execute roddy remote.
Currently, roddy is execute locally. Therefore roddy needs also to be configured for ssh to
connect to the submission host. Also if roddy uses to much memory, it can cause problems for OTP.

Therefore it was decided to execute roddy also remotely. Since then there wouldn't memory problems for
OTP server. Also roddy do not need to be configured for remote access, since it can submit the jobs locally.

Release note:
The roddy call was changed from local to remote call. For the remote call Groovy and Java are needed to be initialized.
For initializing Groovy already a ProcessingOption exists, but not for Java. Therefore the new ProcessingOption COMMAND_ACTIVATION_JAVA
is added. It should contain the command to initialize Java.

TODO:
- init the new ProcessingOption COMMAND_ACTIVATION_JAVA
- change following things in the application.ini
  - adapt the property CLI.executionServiceClass from de.dkfz.roddy.execution.io.SSHExecutionService to de.dkfz.roddy.execution.io.LocalExecutionService
  - the option used for the remote access can be removed. That are:
    - CLI.executionServiceAuth
    - CLI.executionServiceHost
    - CLI.executionServicePasswd
    - CLI.executionServiceStorePassword
    - CLI.executionServiceUseCompression
    - CLI.executionServiceUser