add: OTP-2985: Reactivate the notification if a blocked bam file is accepted.
When the qc traffic light block some bam files, the following analysis are not running.
Therefore the tracking service mark the ticket as finished as soon as all other workflows
of the ilse finished. If the bam files are accepted later, the analysis will run, but it is not
longer tracked by the tracking service, since the otrs ticket are marked as finished.

Since it is wished to get the notification in this case, the code is adapted to reset the notification
in the otrs ticket, so it is tracked again.