chg: Rewrite script load bedfiles
The script was designed to track all loaded bedfiles for reloading.
But that way it is not used. Usually we want to load one bedfile to some reference genomes.

Therefore I change the input to be more inuitive.
But that change break the test for the script, so I removed it.