bugfix: Handle "Resume as" switch in dicom logger onAuthenticationSwitchUserEvent together with backdoor
In development it is possible to use a backdoor instead of login every time. The user switch does work well, but
the 'Resume as' afterwards cause a NullPointerException. The reason therefore is that the target property of the
event is null instead of user principal. And the logger try to get from it user username.