bugfix: Resolve problem of double setting of threadlog during notification
The method ProjectOverviewService.getRoddyProcessOutput() has set the threadlog to log the call
of 'printidlessruntimeconfig'. But the method is also called during the notification job. And
in that case a thread log is already set. And a threadlog can not be set twiced. But since the log
is not really needed for the GUI, I remove the call.