Bugfix: Fix QcTrafficLightCheckService.handleQcCheck to sent notify emails for auto accepted thresholds
In QcTrafficLightCheckService.handleQcCheck the emails are only send for blocked files, not for auto accepted files.
That was changed.

LinkFilesToFinalDestinationServiceTests contains many test for cases, which were moved to qcTrafficLightCheckService.handleQcCheck and
do not need to test anymore. Since they were now broken, I remove that test and keep only one, which I move to Spec.