bugfix: Remove currently handling of qc for analysis
Currently, the result of qc handling of the pipelines were written to bam file. Since we have now next to
PASS and BLOCKKED also AUTO_ACCEPED, that do not work anymore correct. Since we haven't yet any thressholds for
analysis, this bugfix remove qc handling for analysis. To implement a proper way, the jira issue
OTP-3097 was created.