chg: Update alignment workflow for wgbs/wgbs_Tag to 1.2.73-2
The new version support the new roddy option for PBAT:
 <cvalue name="reorderUndirectionalWGBSReadPairs" value="true" type="boolean"/>

If needed, it needs to be set in the configuration file after creation of the configuration.

Please note, it can only set for the complete seqtype, therefore a project may not mix PBAT and not PBAT data.

Release note:
Please update the processing option OptionName.PIPELINE_RODDY_ALIGNMENT_DEFAULT_PLUGIN_VERSION for WGBS and WGBS_TAG
to the new version 1.2.73-2