chg: Adapt OTP to use for SNV, indel, sophia and chipseq alignment the test resource set
Adapt Indel and SNV to request the test resource set for workflow test.

Adapt alignment to request also for chipseq alignment the resource test set. The other alignments already use the test resource sets.

Also create the missing test resource sets in the roddy configs.

On this way, the will use always the devel queue, which is not requested so much.