Bugfix: Correct transaction place
The transaction around in ClusterJobMonitor.handleFinishedClusterJob cause problems, if notifyJobAboutFinishedClusterJob failed,
since then the changed cluster job state is rollbacked, but the monitoring job is removed from the list of running job. Therefor the checker
take that cluster job again, but do not find monitor job for the notification. This cause an endlees loop.