bugfix: Use lock for iteration on list running
Even though asSynchronized() is called on the List, it does not use a
synchronized iterator. (See code of iterator() of
SynchronizedCollection)

Instead it refers to the iterator of the internally held,
unsynchronized, Collection object. This fact is also mentioned as a
comment inside the iterator function.

Because of this we still have to use a lock around our find function.