chg: Adapt Inform identifier
The Inform submitter startet to use the order number to distinct between samples.
But the old implementation in OTP ignore these, so that the are merged. That causes
many da swaps to split them again.

Therefore the parser are changed to include both numbers in the sample type.

Also the check for sample types of existing samples are removed, because that would
make problems during the change of the types.