chg: Add Sample identifier parser for OE0290_EORTC
The project use the hipo schema, but the project is named diferently. Therefore
an adapted parser is needed, which handle the project differently.

Therefore the shared part of Hipo2SampleIdentifierParser was extracted to the abstract base class AbstractHipo2SampleIdentifierParser,
and Hipo2SampleIdentifierParser has obly that which is different to OE0290_EORTC_SampleIdentifierParser