bugfix: Fix validation exception in cross project sample swap
When swapping samples between projects we change the MetaDataEntries
and adapt the project in the DataFiles.
Because of Grails3's automatic deep validation we have to mind the
order in which we do this, otherwise we run into a validation
exception thrown by the custom validator of the DataFiles project
field.

The problem is that the project of the SeqTrack does not match with
the project field but the deep validation on the MetaDataEntry
forces the check on DataFile.

The solution is to first adapt all DataFiles, save them and then
adapt the Samples/SampleIdentifiers/MetaDataEntries.