OTP-3191: Add initial version of README, fit for public consumption.
This is definitely not all that a proper open-source-project
README should be, but it's a start that specifies the bare minimum.
It should provide a skeleton to expand once we get more
documentation out in public.

We borrow some elements from the OTP-paper, which we also point at.
A lot of discussion on terminology and phrasing was already done in
the course of drafting that paper, so lets respect that effort!