Added parser for generic (and simple) sample identifier
This parser can be used to parse samples without a project specific parser. It is designed to provide a generic sample identifier constructed as follows: (PROJECT)(PID)(SAMPLE_TYPE) It can be easily generated and used by both customers and external tools. With this parser, manual intervention can be reduced.