bugfix: OTP-3238: Secure autoImport with secret and create then an authentification object
The import uses now methods, with require the role OPERATOR. But the auto import does not use any
authentification.

Therefore the auto import is now secured by an secret defined in .otp.properties. If the correct secret is given,
an operator authentification object is created for the duration of the import.