bugfix: OTP-3240: Handle double send notification.
When for a ticket the last two workflows finished together so both run simultanly in NotificationCreator.setFinishedTimestampsAndNotify,
and both pass the check for sending the notification, then both workflows send the notification. To avoid this, the method
OtrsTicketService.saveEndTimeIfNeeded return if it set the value in the call or if it was set before. That info is used to avoid
sending the email twiced.