bugfix: otp-187: Fixed "execute roddy, roddy call succeeds, one clusterJob fails, RoddyJob failed and is successfully restarted"
Currently AbstractRoddyJobIntegrationTests."execute roddy, roddy call succeeds, one clusterJob fails, RoddyJob failed and is successfully restarted" does not end until it runs into timeout (60h). This is caused by the LinkGenerator. He needs grails.serverURL to be setm which is not the case for workflow tests. This cases an unhanded Exception to be thrown. This in turn causes ScheduledExecutorService.scheduleWithFixedDelay for ClusterJobMonitor to get canceled. And this in turn causes ClusterJobMonitor.handleFinishedClusterJob to not be executed which prevents the test from noticing that all clusterjobs have finished and thus endlessly waiting for it.

Also decreased the test duration to 30min