Improve calculation of step duration on Process view
The calculation was off, when there were multiple updates with the state
STARTED for one ProcessingStep. In this case it calculated the
difference between the newest STARTED and FINISHED.

This always happens after a shutdown of OTP, as the jobs get a update
with RESUMED and another STARTED.

Now it uses the first update regardless of state. The last update is
only used when it is in a state that is considered finished, otherwise
it uses the current time.