chg: otp-197: Add to SimpleProjectIndividualSampleTypeParser a 4. group for the shown identifier
The SimpleProjectIndividualSampleTypeParser get a 4. group to hold the identifier shown in the otp gui.

The parser is used to transfer the project, pid, sample type information from the guide to otp. But it is not
the original identifier, which should shown to the user. Therefore the original identifier is now additional transfered
in a 4. group to OTP. And OTP use only that part for display.