bugfix: otp-199: Send for every file access click an email to the operators
Currently the code check changes against the file group and send only an email, if it changes against that.
But since changes are done manually, there are a delay on the file system. In case the permission is change
back before it was changed on the file ystem, the operator are not informed about the inverse change, so that
then the change on the file system are changed wrongly. Now OTP send always the email, so that the operator
are also informed about the inverse change.