Bugfix: show actual date and time in timestamps of processing views
The source of the bug was a wrong conversion between the dates passed
from Java to JavaScript and a shift in timezone.

Now I have introduced a separate Java class, which converts Java dates
into the appropriate timestamp and only pass that value to the
javascript. The Javascript now only handles the display, not the
calculation.