otp-123: Shorten table headers on CellRanger QC page
In discussion with a CellRanger user I found out that the all values of
interest can be found in the web_summary.html file, which is the main
result bioinformaticians are looking for.

Because of this I cut down the headers to a maximum of 4 lines and
linked the summary file in the front.