otp-213: Remove "addition" message template
The "addition" templates used for the notification were removed from the
i18n but were still referenced in parts of the code. With a recent
change this caused an exception to be logged.

As there should be no unknown templates, the false use cases were
removed and the function is now stricter with the template names,
throwing the exception and not just logging it.