bugfix: otp-216: Correct show of index read in fastqc column
The column fastqc of the sequence page has assume one or two reads, but no index. With the index the logig was broken.
The reads are all numbered with prefix R, and since the sort logic was broken, it could be any order.

The fix change the Prefix to I for index files and fix the order.
Also it optimize the search for the datafiles for an sequnce and add only the needed infos of datafile to the json.