bugfix: otp-244: Set permission of multi level project path correct
Currently, if a multi level path is created, all levels get the permission 2750, but the
upper level get the group of the project directory. This can cause, that user can not access
their data, because they do not come to their project directory.
Therefore the parant directories should be created with permission 2755.