bugfix: otp-250: Prevent double click bug on ProjectUser GUI
When rapidly clicking the enable/disable submit button on the user
management GUI, it submitted the toggle command multiple times. This
cause the function to execute multiple times, sending mails, setting
attributes.

Now the toggle functions have been refactored into straight setters and
the submit button is disabled until the submit is finished and the page
is reloaded.

Also removed the overridden getter for manageUser and adapted affected
pieces.