otp-210: Allow configuration of ReferenceGenome per CellRanger execution
The reference genome index is now configured manually via the GUI on the
cell ranger execution page.

New results will have the reference genome index as part of the
directroy name.

The GUI has been adapted to show the index per work package and no
longer shows the complete toString of the CRMWP.