chg: otp-334: Add additional case where ad group was changed not on request of OTP
Currently OTP assumes, that all group changes are done on request of OTP. But there are cases,
where groups are changed in ldap not on request of OTP. In that case OTP had displayed, that the
operator will do the changes soon, but they do not know about it and therefore do not do that.

To avoid that confusing in the future, OTP knows now if a change was requested or not. If the change was requested,
OTP display that there was a change requested and the operator know about it and will do it soon.
Otherwise OTP display a question marc to indicate, that there is an mismatch.