new: otp:annotation for quick and consistent annotations on the GUI
A new component in the OtpTagLib which serves as a consistent way to
provide different types of annotations to the user. It consists of a
simple div with a strong border and a light background. There are
predefined subtypes covering infos, warnings and erros. A neutral type
serves as a default.