otp-207 [1]: new feature: opt-out mechanism for searchable dropdowns.
In preparation for default select2 usage, provide an opt-out.
The current implementation looks a bit silly, but will make more sense once
we do the planned switch to "all selects are searchable by default"

There are situations where select2 needs some special attention, so we
provide the 'dont-use-select-2' CSS-class to mark that we thought about it
but don't want it to use select2 (yet).
Please document with a comment why you add this somewhere.