Refactor: Deduplicate top-of-page filter/search CSS.
We have a lot of "nice rounded box at top of page" on different pages.
The CSS for these was copy-pasted everywhere, in subtly different versions.
Here we make a re-usable snippet for all of them.

The pre-existing snippets where re-used "creatively" by setting illogical
IDs ("projectsGroupbox") here and there. This is also replaced with proper
class-usage.

Since 'linear-gradient' was standardised in 2012, with support in
all browsers, do away with the -moz-/-webkit- prefixes for that.

We also use this new header for the statistics pages, eliminating their
custom style.