chg: otp-345: Estimate and fill in timestamp fields of Entity
This commit contains a file with multiple scripts, which allow filling
in timestamps of domain objects based on their IDs.

Normally this is not needed, as each object has its timestamps
automatically maintained by grails via the Entity trait. However this
has only be added with a commit in Mai 2019, when some instances of OTP
were already running.

This script serves to estimate the timestamps of objects created before
the mentioned commit, which is done based on the (almost) global unique
ID and some entities which always had a timestamp like 'DataFile'.