Style: New OTP-themed Favicon!
OTP deserves it's own favicon, not a copy of the www.dkfz.de icon.

Additionaly, this change improves feedback during development and testing,
by styling the favicon depending on the servers' environment variable.

During development it can easily happen that we are working on multiple
deployments, e.g. veryfying something in "production", testing a new
feature on "gamma", and debugging something locally.
To help us keep track of which is which, we already display the environment
name in the header bar.

By default we use only the ones based on environment:
  - 'production' gets OTP-blue
  - 'development' a 'pale green'
  - alpha/beta/gamma get 'light goldenrod'

Local overrides (e.g. our "instance annotator" greasemonkey extension) can also use:
  - 'tomato'
  - 'light slate blue'