otp-207 [6]: Force Searchable Dropdowns into OTP-styling.
This is a bit of a hacky solution, but the "proper" alternative is to
write a select2 theme file from scratch, which is fairly heavy overkill
to change two colours.

Since the 'computedwidth' setting doesn't always guess right, provide
some overrides in the form of explicit widths.
(It seems the width calculation doesn't take the no-selection placeholder
into account)
  * Cellranger/index
  * QC thresholds (both default and project-specific)
  * project document filetypes

To enable the custom-CSS feature of Select2, we need to use the 'full'
version, which fortunately is also part of the WebJar-distribution.